Disable Coevery.Relationship PostBuildEvent
ClickToBuild.cmd build fails due to a post-build event to copy the Coevery.Relationship.dll file to
the Coevery.DynamicTypes bin folder set in the Coevery.Relationship project:

"$(TargetPath)" "$(SolutionDir)Orchard.Web\Modules\Coevery.DynamicTypes\bin\"

This fails because a bin sub-folder is not yet created in the Coevery.DynamicTypes project folder. Manually
creating this folder and running ClickToBuild.cmd works around the issue.

However, I do not believe it is required and it is not present in the current Covery repository develop branch.